feat(meta): Sourcehut -> GitHub
Unfortunately, moving this project back to a GitHub repo, and Render as
a host. I'm doing this reluctantly, because I like Sourcehut and its
tooling, but Sourcehut pages blocks external resources.

This does not allow me to embed for example YouTube videos, but also
breaks the newsletter signup form.

Hopefully this is a temporary reversal.